use flat structure in types/, rename common to internal
