+++ packages/joint-core/types/anchors.d.ts
@@ -1,5 +1,5 @@
 import type * as dia from './dia';
-import type * as g from './g';
+import type * as g from './geometry';
 import type * as linkAnchors from './linkAnchors';
 
 export interface ElementAnchorArguments {
