+++ packages/joint-core/types/attributes.d.ts
