+++ packages/joint-core/types/connectionStrategies.d.ts
