+++ packages/joint-core/types/connectors.d.ts
@@ -1,5 +1,5 @@
 import type * as dia from './dia';
-import type * as g from './g';
+import type * as g from './geometry';
 
 export interface NormalConnectorArguments {
     raw?: boolean;
