+++ packages/joint-core/types/index.d.ts
@@ -4,24 +4,24 @@ export {
     setTheme,
     version
 } from './joint';
-export * as shapes from './namespaces/shapes';
-export * as attributes from './namespaces/attributes';
-export * as dia from './namespaces/dia';
-export * as elementTools from './namespaces/elementTools';
-export * as linkTools from './namespaces/linkTools';
-export * as anchors from './namespaces/anchors';
-export * as linkAnchors from './namespaces/linkAnchors';
-export * as highlighters from './namespaces/highlighters';
-export * as layout from './namespaces/layout';
-export * as connectionPoints from './namespaces/connectionPoints';
-export * as connectionStrategies from './namespaces/connectionStrategies';
-export * as connectors from './namespaces/connectors';
-export * as routers from './namespaces/routers';
-export * as mvc from './namespaces/mvc';
-export * as util from './namespaces/util';
+export * as shapes from './shapes';
+export * as attributes from './attributes';
+export * as dia from './dia';
+export * as elementTools from './elementTools';
+export * as linkTools from './linkTools';
+export * as anchors from './anchors';
+export * as linkAnchors from './linkAnchors';
+export * as highlighters from './highlighters';
+export * as layout from './layout';
+export * as connectionPoints from './connectionPoints';
+export * as connectionStrategies from './connectionStrategies';
+export * as connectors from './connectors';
+export * as routers from './routers';
+export * as mvc from './mvc';
+export * as util from './util';
 export {
     V,
     Vectorizer,
     VElement
-} from './namespaces/V';
-export * as g from './namespaces/g';
+} from './vectorizer';
+export * as g from './geometry';
