+++ packages/joint-core/types/joint.d.ts
@@ -1,4 +1,4 @@
-import type * as util from './namespaces/util';
+import type * as util from './util';
 
 export namespace config {
     var useCSSSelectors: boolean;
@@ -19,18 +19,18 @@ export function setTheme(theme: string): void;
 
 export const version: string;
 
-export * as shapes from './namespaces/shapes';
-export * as attributes from './namespaces/attributes';
-export * as dia from './namespaces/dia';
-export * as elementTools from './namespaces/elementTools';
-export * as linkTools from './namespaces/linkTools';
-export * as anchors from './namespaces/anchors';
-export * as linkAnchors from './namespaces/linkAnchors';
-export * as highlighters from './namespaces/highlighters';
-export * as layout from './namespaces/layout';
-export * as connectionPoints from './namespaces/connectionPoints';
-export * as connectionStrategies from './namespaces/connectionStrategies';
-export * as connectors from './namespaces/connectors';
-export * as routers from './namespaces/routers';
-export * as mvc from './namespaces/mvc';
-export * as util from './namespaces/util';
+export * as shapes from './shapes';
+export * as attributes from './attributes';
+export * as dia from './dia';
+export * as elementTools from './elementTools';
+export * as linkTools from './linkTools';
+export * as anchors from './anchors';
+export * as linkAnchors from './linkAnchors';
+export * as highlighters from './highlighters';
+export * as layout from './layout';
+export * as connectionPoints from './connectionPoints';
+export * as connectionStrategies from './connectionStrategies';
+export * as connectors from './connectors';
+export * as routers from './routers';
+export * as mvc from './mvc';
+export * as util from './util';
