+++ packages/joint-core/types/linkAnchors.d.ts
